src: linux: fix gcc conversion issue
../src/handmadehero_linux.c: In function ‘joystick_key’:
../src/handmadehero_linux.c:331:33: error: conversion from ‘int’ to ‘unsigned char:1’ may change value [-Werror=conversion]
  331 |     controller->start.pressed = ((u8)value) & 1;
      |                                 ^
../src/handmadehero_linux.c:335:32: error: conversion from ‘int’ to ‘unsigned char:1’ may change value [-Werror=conversion]
  335 |     controller->back.pressed = ((u8)value) & 1;
      |                                ^

+++ src/handmadehero_linux.c
@@ -328,11 +328,11 @@ static void joystick_key(struct linux_state *state, u16 type, u16 code,
   } break;
 
   case BTN_START: {
-    controller->start.pressed = value & 1;
+    controller->start.pressed = (u8)(value & 1);
   } break;
 
   case BTN_SELECT: {
-    controller->back.pressed = value & 1;
+    controller->back.pressed = (u8)(value & 1);
   } break;
 
     /* normalize stick x movement */
